requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pillow
python-multipart
timm
